Move find_and_replace.py into browser/dialogs
