+++ qt/aqt/browser/__init__.py
@@ -1,10 +1,13 @@
 # Copyright: Ankitects Pty Ltd and contributors
 # License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-
 from __future__ import annotations
 
+import sys
+
+import aqt
+
 from .browser import Browser
-from .dialogs import CardInfoDialog, ChangeModel, FindDupesDialog
+from .dialogs import CardInfoDialog, ChangeModel, FindAndReplaceDialog, FindDupesDialog
 from .sidebar import (
     SidebarItem,
     SidebarItemType,
@@ -30,3 +33,7 @@ from .table import (
     StatusDelegate,
     Table,
 )
+
+# aliases for legacy pathnames
+sys.modules["aqt.find_and_replace"] = sys.modules["aqt.browser.dialogs"]
+aqt.find_and_replace = sys.modules["aqt.browser.dialogs"]  # type: ignore
